RichMenuListContentのprop・型対応
