+++ src/components/molecules/RichMenulist/type.ts
@@ -1,2 +1,8 @@
-export type RichMenulistProps = {
-}
+import { ReactElement } from "react";
+import { ListItemProps } from "@components/atoms/ListItem/type";
+
+export type RichMenuListProps = {
+  title: string;
+  titleIcon?: ReactElement;
+  items: ListItemProps[];
+};
