+++ src/components/organisms/RichMenu/RichMenuContents/RichMenuContents.tsx
@@ -3,9 +3,16 @@ import { RichMenuListContent } from "../RichMenuListContent";
 import type { RichMenuContentsProps } from "./type";
 
 export const RichMenuContents = ({ richMenuType }: RichMenuContentsProps) => {
-  if (richMenuType === "description") return <RichMenuDescriptionContent />;
+  if (richMenuType === "description")
+    return (
+      <RichMenuDescriptionContent
+        headingText={""}
+        description={""}
+        imgSrc={""}
+      />
+    );
 
-  if (richMenuType === "list") return <RichMenuListContent />;
+  if (richMenuType === "list") return <RichMenuListContent listGroup={[]} />;
 
   return <></>;
 };
