+++ src/components/organisms/RichMenu/RichMenuListContent/type.ts
@@ -1,2 +1,5 @@
+import { RichMenuListProps } from "@components/molecules/RichMenulist/type";
+
 export type RichMenuListContentProps = {
-}
+  listGroup: RichMenuListProps[];
+};
